requirements.txt
Django==1.7.4
Markdown==2.5.2
Pillow==2.7.0
django-cors-headers==1.0.0
django-filter==0.9.2
djangorestframework==3.0.4
wsgiref==0.1.2
